20221028 숙소 필터 및 카페 필터 완

+++ Petkage/src/main/java/com/finalproject/petkage/wherego/model/vo/avgReview.java
@@ -11,7 +11,7 @@ public class avgReview {
 
 	private String spotName;
 	
-	private double reviewAvg;
+	private int reviewAvg;
